Fix a couple bugs in "position" command handling.
Bug #1: If position command was issued before isready then resulting position would get discarded by next command that caused engine to be initialized.  Issuing the position command before the isready command is non-standard so this was not a major bug.

Bug #2: If the "moves" keyword was used in the position command it would not process the moves.  This was a major bug.  You could trigger it in analysis mode in Chess Arena whenever you use the back arrow.